refactor(frontend): extract OpenAI session/WebRTC logic to useOpenAISession hook (#14)
