+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useRef, useState } from "react";
+import { useState } from "react";
 import {
   Button,
   Flex,
@@ -12,25 +12,10 @@ import {
 
 const INITIAL_COUNT = 0;
 
-type OpenAISessionResponse = {
-  client_secret: {
-    value: string;
-    [key: string]: unknown;
-  };
-  [key: string]: unknown;
-};
-
-const fetchOpenAISession = async (): Promise<OpenAISessionResponse> => {
-  const res = await fetch("/api/openai/session");
-  if (!res.ok) {
-    throw new Error("Failed to fetch OpenAI session: " + res.statusText);
-  }
-  return await res.json();
-};
-
 import EventList from "./EventList";
 import { useEventForm } from "./hooks/useEventForm";
 import EventFormDialog from "./components/EventFormDialog";
+import { useOpenAISession } from "./hooks/useOpenAISession";
 
 type Event = {
   id: number;
@@ -40,17 +25,9 @@ type Event = {
   end_time: string;
 };
 
-
 const App: React.FC = () => {
   const [count, setCount] = useState<number>(INITIAL_COUNT);
   const [name, setName] = useState<string>("");
-  const [isConversing, setIsConversing] = useState<boolean>(false);
-  const [sessionToken, setSessionToken] = useState<string | undefined>(undefined);
-  const [error, setError] = useState<string | undefined>(undefined);
-  const [peerConnection, setPeerConnection] = useState<RTCPeerConnection | null>(null);
-  const [localStream, setLocalStream] = useState<MediaStream | null>(null);
-  const audioRef = useRef<HTMLAudioElement>(null);
-
   // Event form and dialog state/logic
   const {
     eventForm,
@@ -61,10 +38,33 @@ const App: React.FC = () => {
     resetEventForm,
   } = useEventForm();
 
+  // OpenAI session and conversation logic
+  const {
+    isConversing,
+    sessionToken,
+    error,
+    audioRef,
+    handleStartConversation,
+    handleStopConversation,
+  } = useOpenAISession((args) => {
+    // Prefill event form when function_call is received
+    openEventDialog({
+      title: args.title || "",
+      description: args.description || "",
+      startDate: args.start_time ? args.start_time.slice(0, 10) : "",
+      startTime: args.start_time ? args.start_time.slice(11, 16) : "",
+      endDate: args.end_time ? args.end_time.slice(0, 10) : "",
+      endTime: args.end_time ? args.end_time.slice(11, 16) : "",
+    });
+  });
+
   // State for events list
   const [events, setEvents] = useState<Event[]>([]);
   const [eventsLoading, setEventsLoading] = useState<boolean>(true);
   const [eventsError, setEventsError] = useState<string | undefined>(undefined);
+
+  // Local state for event form validation errors
+  const [formError, setFormError] = useState<string | undefined>(undefined);
 
   // Fetch events from backend
   const fetchEvents = async () => {
@@ -91,158 +91,6 @@ const App: React.FC = () => {
     fetchEvents();
   }, []);
 
-  // Establish WebRTC connection to OpenAI Realtime API
-  const handleStartConversation = async (): Promise<void> => {
-    setError(undefined);
-    setSessionToken(undefined);
-    setPeerConnection(null);
-    setLocalStream(null);
-    try {
-      const data = await fetchOpenAISession();
-      let token: string | undefined = undefined;
-      if (
-        data &&
-        typeof data === "object" &&
-        "client_secret" in data &&
-        data.client_secret &&
-        typeof data.client_secret === "object" &&
-        "value" in data.client_secret
-      ) {
-        token = data.client_secret.value;
-      }
-      setSessionToken(token);
-
-      if (!token) throw new Error("No ephemeral session token received");
-
-      // 1. Create peer connection
-      const pc = new RTCPeerConnection();
-
-      // 2. Set up to play remote audio from the model
-      pc.ontrack = (e) => {
-        if (audioRef.current) {
-          audioRef.current.srcObject = e.streams[0];
-        }
-      };
-
-      // 3. Add local audio track for microphone input
-      const ms = await navigator.mediaDevices.getUserMedia({ audio: true });
-      setLocalStream(ms);
-      pc.addTrack(ms.getTracks()[0]);
-
-      // 4. Set up data channel for events (function calling, etc.)
-      const dc = pc.createDataChannel("oai-events");
-
-      // Function schema for calendar event creation
-      const calendarEventFunction = {
-        type: "function",
-        name: "create_calendar_event",
-        description: "Create a calendar event with structured data.",
-        parameters: {
-          type: "object",
-          properties: {
-            title: { type: "string", description: "Event title" },
-            description: { type: "string", description: "Event description" },
-            start_time: { type: "string", description: "Event start time (ISO 8601)" },
-            end_time: { type: "string", description: "Event end time (ISO 8601)" }
-          },
-          required: ["title", "start_time", "end_time"]
-        }
-      };
-
-      // Prefill event form when function_call is received
-      dc.addEventListener("message", (e) => {
-        try {
-          const data = typeof e.data === "string" ? JSON.parse(e.data) : e.data;
-          // Listen for response.done with function_call output
-          if (
-            data?.type === "response.done" &&
-            Array.isArray(data.response?.output)
-          ) {
-            for (const item of data.response.output) {
-              if (
-                item.type === "function_call" &&
-                item.name === "create_calendar_event" &&
-                typeof item.arguments === "string"
-              ) {
-                const args = JSON.parse(item.arguments);
-                console.log({args})
-                // Prefill event form state (if available in args)
-                openEventDialog({
-                  title: args.title || "",
-                  description: args.description || "",
-                  startDate: args.start_time ? args.start_time.slice(0, 10) : "",
-                  startTime: args.start_time ? args.start_time.slice(11, 16) : "",
-                  endDate: args.end_time ? args.end_time.slice(0, 10) : "",
-                  endTime: args.end_time ? args.end_time.slice(11, 16) : "",
-                });
-              }
-            }
-          }
-        } catch {
-          // Ignore parse errors
-        }
-      });
-
-      // After connection, send session.update to register the function
-      dc.addEventListener("open", () => {
-        const sessionUpdate = {
-          type: "session.update",
-          session: {
-            tools: [calendarEventFunction],
-            tool_choice: "auto"
-          }
-        };
-        dc.send(JSON.stringify(sessionUpdate));
-      });
-
-      // 5. Start the session using SDP
-      const offer = await pc.createOffer();
-      await pc.setLocalDescription(offer);
-
-      const baseUrl = "https://api.openai.com/v1/realtime";
-      const model = "gpt-4o-realtime-preview-2024-12-17";
-      const sdpResponse = await fetch(`${baseUrl}?model=${model}`, {
-        method: "POST",
-        body: offer.sdp,
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/sdp",
-        },
-      });
-
-      const answer: RTCSessionDescriptionInit = {
-        type: "answer",
-        sdp: await sdpResponse.text(),
-      };
-      await pc.setRemoteDescription(answer);
-
-      setPeerConnection(pc);
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        setError(error.message);
-      } else {
-        setError("Unknown error");
-      }
-    }
-    setIsConversing(true);
-  };
-
-  // Clean up WebRTC connection and audio streams
-  const handleStopConversation = (): void => {
-    setIsConversing(false);
-    if (peerConnection) {
-      peerConnection.close();
-      setPeerConnection(null);
-    }
-    if (localStream) {
-      localStream.getTracks().forEach((track) => track.stop());
-      setLocalStream(null);
-    }
-    if (audioRef.current) {
-      audioRef.current.srcObject = null;
-    }
-  };
-
   let conversationButton;
   if (isConversing) {
     conversationButton = (
@@ -258,7 +106,6 @@ const App: React.FC = () => {
     );
   }
 
-
   // Save event to backend
   const handleSaveEvent = async () => {
     // Basic validation
@@ -269,9 +116,10 @@ const App: React.FC = () => {
       !eventForm.endDate ||
       !eventForm.endTime
     ) {
-      setError("Please fill in all required fields.");
+      setFormError("Please fill in all required fields.");
       return;
     }
+    setFormError(undefined);
     // Combine date and time into ISO 8601 strings
     const start_time = `${eventForm.startDate}T${eventForm.startTime}:00`;
     const end_time = `${eventForm.endDate}T${eventForm.endTime}:00`;
@@ -288,16 +136,16 @@ const App: React.FC = () => {
         }),
       });
       if (!res.ok) {
-        throw new Error("Failed to create event: " + res.statusText);
+        setFormError("Failed to create event: " + res.statusText);
+        return;
       }
       // Optionally, you could use the returned event here
       closeEventDialog();
       resetEventForm();
-      setError(undefined);
       // Refresh events list after saving
       fetchEvents();
     } catch (err) {
-      setError(
+      setFormError(
         err instanceof Error ? err.message : "Unknown error creating event"
       );
     }
@@ -351,7 +199,7 @@ const App: React.FC = () => {
             onChange={handleEventFormChange}
             onSave={handleSaveEvent}
             onClose={closeEventDialog}
-            error={error}
+            error={formError}
           />
         </Flex>
         <EventList events={events} loading={eventsLoading} error={eventsError} />
